refactor(core/server): extract route specification types into ./server/route-spec.ts
Moves the family of route spec types out of `server.ts`:

- `SchemaDefinition`, `WithOpenAPI`, `RouteSpec`
- `EffectiveSchemas<T>` (internal lookup helper, now exported for
  `HandlerContext`)
- `InferResponseType`, `InferResponses`, `ValidResponse`
- `ResponseForStatus`, `ValidStatusCodes`, `InferResponse`

`HandlerContext<T>` stays in `server.ts` because its `response` field
is typed against `ResponseTools<T>`, which is tightly coupled to the
response helpers in that file — splitting it out would create a
circular import. It now consumes the shared `EffectiveSchemas<T>`
helper from route-spec.ts.

Public type surface is re-exported from `server.ts` so consumers that
imported these names from `@ts-api-kit/core/server` keep working.

`wc -l`: server.ts 1641 → 1550; route-spec.ts is 136 lines. Tests,
lint, and type-check remain green.
